bookstore-micro05/staff-service/requirements.txt
Django>=5.0
djangorestframework
mysqlclient
gunicorn
cryptography